Move old message definitions to deprecated ns
