+++ src/Messages/Deprecated/Generic.cs
@@ -1,0 +1,7 @@
+namespace EulynxLive.Messages.Deprecated
+{
+    public enum PdiVersionCheckResult : byte {
+        NoMatch = 0x1,
+        Match = 0x2
+    }
+}
