--- src/Messages/Generic.cs
@@ -1,7 +1,0 @@
-namespace EulynxLive.Messages
-{
-    public enum PdiVersionCheckResult : byte {
-        NoMatch = 0x1,
-        Match = 0x2
-    }
-}
